refactor(types): standardize file naming convention to .types.ts suffix
